backend/requirements.txt

fastapi
uvicorn
websockets
openai
python-dotenv
streamlit
streamlit-chat
requests
asyncio-mqtt
aiohttp
